Shuffling linelist data fetching to an API endpoint on curator-service.
