+++ verification/curator-service/api/src/index.ts
@@ -6,6 +6,7 @@ import axios from 'axios';
 import validateEnv from './util/validate-env';
 
 // Controllers (route handlers).
+import * as casesController from './controllers/cases';
 import * as homeController from './controllers/home';
 import * as sourcesController from './controllers/sources';
 
@@ -24,6 +25,7 @@ app.get('/', homeController.index);
 
 // Configure API app routes.
 const apiRouter = express.Router();
+apiRouter.get('/cases/', casesController.list);
 apiRouter.get('/sources/', sourcesController.list);
 apiRouter.get('/sources/:id([a-z0-9]{24})', sourcesController.get);
 apiRouter.post('/sources/', sourcesController.create);
