+++ verification/curator-service/api/test/home.test.ts
@@ -1,35 +1,8 @@
 import request from 'supertest';
 import app from '../src/index';
-import axios, { AxiosResponse } from 'axios';
-
-jest.mock('axios');
-const mockedAxios = axios as jest.Mocked<typeof axios>;
 
 describe('GET /', () => {
-    it('should return 200 OK with latest case date', (done) => {
-        const cases = [
-            {
-                _id: 'abc123',
-                outcome: 'recovered',
-                date: new Date('2020-04-21').toJSON(),
-            },
-            {
-                _id: 'def456',
-                outcome: 'recovered',
-                date: new Date('2020-04-22').toJSON(),
-            },
-        ];
-        const axiosResponse = {
-            data: cases,
-            status: 200,
-            statusText: 'OK',
-            config: {},
-            headers: {},
-        };
-        mockedAxios.get.mockResolvedValue(axiosResponse);
-
-        request(app)
-            .get('/')
-            .expect(200, /2020-04-22/, done);
+    it('should return 200 OK', (done) => {
+        request(app).get('/').expect(200, done);
     });
 });
